[update] 更新NVOD视图及NVOD业务处理逻辑；优化流密度计算方法。

+++ modules/assistant-api/src/main/java/m2tk/assistant/api/domain/StreamDensityBulk.java
@@ -28,7 +28,7 @@ public class StreamDensityBulk
     private long startPosition;
     private long maxDensity;
     private long minDensity;
-    private long avgDensity;
+    private double avgDensity;
 
     public final int[] getDensities()
     {
