+++ modules/assistant-app/src/main/java/m2tk/assistant/app/kernel/entity/DensityStatViewEntity.java
@@ -29,7 +29,7 @@ public class DensityStatViewEntity
     @TableField("`count`")
     private Integer count;
     @TableField("avg_density")
-    private Long avgDensity;
+    private Double avgDensity;
     @TableField("min_density")
     private Long minDensity;
     @TableField("max_density")
